requirements_.txt
altair==4.2.0
boto3++boto3
fastparquet==0.8.1
folium==0.12.1.post1
geopandas==0.11.0
geopy==2.2.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.1
Pillow==9.1.1
plotly==5.9.0
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.10.0
streamlit_folium==0.6.14
streamlit_option_menu==0.3.2
